Add testing flag to PlaygroundParser.
Previous commit suppressed ParsingError exceptions, (the message was
still propagated). The existing tests relied on those exceptions to
test what the parsing should not be able to parse.

Added a flag to PlaygroundParser that can be enabled when testing
to cause PlaygroundParser to re-raise the exception so pytest can
use it.