Fix parsing bug related to parsing name expr.
If a statement contained ONLY a name, example:

foo;

And immediately following it was an assignment:

foo;
bar = 5;

The parser would see the assign op as lookahead token 4 and then try to
parse "foo;" as an assignment statement.

Fixed by checking for existence of semi-colon at lookahead token 2 in
the conditional for assignment statements.