Fix scope bug preventing import system from working
In method '_statements()' a new scope would always be pushed.
This is because this method is used to handle block statements, but it
is also called from '_program()' when parsing the top-level global
statements.

This caused all global symbols to actually be in a child of the global
scope, which was the same for imported symbols.

Added an optional arg 'push_scope' to '_statements()' so that pushing a
new scope is optional. If a block statement is not being handeled,
instead a module is being loaded, or the main program is being run,
don't push a new scope, use the global scope like intended.