Fix subtle assignment bug.
If a variable name was defined and given a value like:

	foo = 5;

and then a function with the same name was defined later:

	def foo(){}

this would cause an exception to be raised, because of how
functions are loaded into the symbol table.

Fixed by checking if the name used for defining the function is already
in the symbol table, and if it is, and if it does not reference a
function dictionary, then reassign that name to a function dict.